오류 수정(delete, upate pin에서 delete contents 같이 해결), 내 핀보기에서 추천 여부 반환

+++ src/main/java/com/seoulapp/withmap/dao/PinDao.java
@@ -6,7 +6,7 @@ import com.seoulapp.withmap.model.Pin;
 
 public interface PinDao {
 
-	List<Pin> getPins(final double latitude, final double longitude, final int radius);
+	List<Pin> getPins(final double latitude, final double longitude, final double radius);
 	List<Pin> getPins(final int userId);
 	Pin get(final int id);
 	int insert(final Pin pin);
